core: db scripts: reduce warnings when creating db
We now change the directory to /  before calling the
database command line tools which suppresses warnings like
'could not change directory to "/root"'

In bareos-config-lib, we now change the log text when it is not
possible to get the parameters via db_check from the configuration from
"Warning" to "Info", and added also the hint that this is usually not an
error.

This should reduce the irritation that usually is created when calling
the database creation scripts.